deepfake_backend/requirements.txt
torch
torchvision
transformers
pillow
fastapi
uvicorn
python-multipart
opencv-python
mtcnn
numpy
tensorflow